this is referencing the previous commit, ii'm still getting the hang of this commit with file message thing lol
created backend functionality for delete and edit. for edit, i just get the
date by id, iterate through the keys of the provided object, and use setattr to
iterate through them.

the only thing here that could  really  use a change is the fact that 'request'
is added as a prop to the object, this should be completely decoupled from the
object data and it should be tied to teh instances of the react functional
components with state. this is a todo! when refactoring the codebase.