that was a long and painful debugging session
that was a huge pain in the ass. essentially, the proxy_pass for / was
working fine. but it'd make requests for static because that's where the
js and css files were, but then nginx would pick that up lol. and so
nginx was like ok yeah you want static i'll give you static, but
i didn't really any logic as to why the static directive was there, it
just kinda was, probably from the previous config. so it would pick that
up and there was a root directive inside it that was some path to the
htdocs on the apache server but that obvisouly wasn't helping

what would have worked was if i had the /static proxy pass to the apache
static, but obviously taht'd be silly because it's already on the
browser and theres no need to re-request static. but that's the only way
it'd work with a static directive, unless there's a way to
self-reference or say "never mind, do what you'd usually do".