slight progress with GCP postgresql
locally, can run the api, file server, and proxy together, with the api
connected to GCP postgresql. there were only a few things to fix, some
slight changes in the config.ini file, some changes to the
requirements.text, and running the container with a mount to the ADC
token. next to do some research on GKE and finish reading the article.
i'm going to see if i can uhh deploy this with apache and nginx proxy
instead of the way he's doing it. that'll also force me to learn more
about how it works