done with file server api call environment variables!
the next step is to use environment variables in the server so that it's
using a local sqlite database during development but the google
postgresql database in production. that'll be tomorrow morning, then
i'll clean up everything one last time, and i'll add some documentation.
then i'll resume development on the weekend