added firebase project to my google cloud project
successfully connected to firebase but that's it. spent a lot of time
trying to get the import working because it was broken for a while and
then it started magically working... not sure why. didn't get as much
done for auth as i wanted but at least i now have the api key for my
project and the front-end app is registered on the firebase project.
Next step is to test out creating and authenticating users, and then
maybe build a UI around it. After that's completely finished, i'll move
on to implementing firebase in the backend