firebaseui is so painful
got firebase authentication , user creation and sign up workign (tested
at lest). also got a decent firebaseui working. was a huge pain in the
butt and chatgpt actually saved the day. i don't understand enough about
how the auth object works so i think i need to read the documentation.
next step is to see if firebaseui can provide everything that i want
(user creation, sign up, sign in with google, etc... and routing to
success page / home page etc... if it doesn't seem capable, i'm going to
have to roll my own. either way, i'm going to have to start thinking
about the sign in flow and maybe build a temporary homepage... but the
main priority right now is getting the UI buitl and functioning for sign
on, and it's probably going to take some time to figure out if
firebaseui is going to work. but if it takes me 2 hours of messing with
firebaseui to realize that it'll work, its worth it because rolling my
own will suck