figured out firebaseui sign up haha
so it turns out that sign up is already implemented as part of the sign
in process. so it seems like i'm going to go with firebaseui at this
stage. this means that tomorrow i'll be doing the sign in flow and
changing things around quite a bit, so start with sign in, then go to
user home or user profile or something like that.. then after that's all
done, i'll go ahead and move towards setting up firebase auth in the
backend. then deployment then feed then deployment then refactoring the
front end yada yada yada