i was too brief on that last commit lol
that was actually a pretty important step in this process since now we
have some basic homepage functionality. it was a little tricky.
i initially thought that if the sign-up page takes me to the
'signInSuccessful' link, then obviously the sign in was successful and
the user will exist. when using auth.currentUser, it returns null when
the auth object is not done intializing (darn async programming) so we have to
use a listener, called onAuthStateChanged, which we can use to change
the state of the functional component using the useState hook