implemented a verifier on the API
two important contributions: intalled and got the admin SDK to work on
the api server, and got an API token verified on the API server. This
laid some groundwork for what will come tomorrow which will be
implementing actual endpoints for uploading data from the server, now
that we can transmit tokens. note to fix the bearer bug first thing.