add middleware to API to verify user tokens, simple test of put/get
token verification works on the API now, and posting and getting a date
from the server is now supported. (not in the front end yet, though)

at this point, a huge overhaul in code maintenance is due. At least
2 days, one for fixing things up and documenting them, and another for
  building nice abstractions around the stuff i've been using so that
when it comes time to build ontop of this functionality, i have some
good tools to work with. So clean up -> document -> build some utilities