add basic but very buggy account crud functionality
Things are getting closer but there's some real issues with the front
end not displaying the most up to date information. The most urgent work
is getting the homepage to display things properly after delete and
edit, and then to clean up the interactions between the APIService and
the server, as there is minimal error handling etc.